# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pds-backend/api/transferType": {
            "get": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Transfer Type List",
                "description": "Total Transfer Type List",
                "operationId": "transferType",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/addTransferType": {
            "post": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Transfer Type List",
                "description": "Add New Transfer Type",
                "operationId": "addTransferType",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added Transfer Type Setup Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/updateTransferType/{id}": {
            "put": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Transfer Type List",
                "operationId": "updateTransferType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "title": {
                                        "type": "text",
                                        "example": "General"
                                    },
                                    "status": {
                                        "type": "text",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer Type Setup Update Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/deleteTransferType/{id}": {
            "delete": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Delete TransferType Type Record",
                "description": "Delete TransferType Type Record With Valid ID",
                "operationId": "deleteTransferType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully, Delete Transfer Type Record"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/activeTransferType/{id}": {
            "patch": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Active TransferType Type Record",
                "description": "Active Specific TransferType Type Record With Valid ID",
                "operationId": "activeTransferType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully, Active Transfer Type Record"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/inactiveTransferTypeRecord/{id}": {
            "patch": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "In-active Transfer Type Record",
                "description": "In-active Specific Transfer Type Record With Valid ID",
                "operationId": "inactiveTransferTypeRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully, In-active Transfer Type Record"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/specificTransferType/{id}": {
            "get": {
                "tags": [
                    "PDS Transfer Type Setup"
                ],
                "summary": "Specific Transfer Type Info",
                "description": "Specific Transfer Type Info",
                "operationId": "specificTransferType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}